+++ CHANGES.rst
@@ -4,6 +4,15 @@ Changes
 2.0.0 (unreleased)
 ------------------
 
+- Improve ``__len__`` function to use underlying dict implementation directly.
+  [rnix]
+
+- Deprecate ``_odict``. user ``_base_odict`` instead.
+  [rnix]
+
+- Rename ``_dict_impl`` to ``_dict_cls`` and ``_list_factory`` to ``_list_cls``.
+  [rnix]
+
 - Refactor package layout to use ``pyproject.toml`` and implicit namespace packages.
   [rnix]